remember which objects were created
the old code got a list of ALL objects (i.e. not rpi2 only) just to check if states/devices are already created. That is a bit of overkill (especially memory wise). On the other hand it did not remember if a state/device was created during runtime, the object list was only updated on startup. 

Now it maintains a map that only stores "true" for all objects created (in previous executions or during runtime) -> sufficient for the check if a device/object needs to be created.